requirements.txt
fastapi
uvicorn[standard]
playwright
gspread
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
requests
